pyproject.toml
[tool.poetry]
name = "html_summary"
version = "0.1.0"
description = ""
authors = ["narumi <toucans-cutouts0f@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.17"
langchain-openai = "^0.1.6"
langchain-text-splitters = "^0.0.1"
python-dotenv = "^1.0.1"
click = "^8.1.7"
loguru = "^0.7.2"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
ruff = "^0.1.8"
toml = "^0.10.2"
mypy = "^1.8.0"

[tool.poetry.scripts]
html-summary = "html_summary.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "C", # flake8-comprehensions
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "W", # pycodestyle warnings

]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
